backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
librosa
numpy
mido
python-dotenv
supabase>=2.0.0
scipy
